// src/types/types.ts
  export type TransactionType =
    | "CustomerPayBillOnline"
    | "CustomerBuyGoodsOnline";
  
  /**
   * The Mobile Number to receive the STK Pin Prompt. This number can be the same as PartyA value above.
   * @type {string}
   * @example "2547XXXXXXXX"
   */
  export type PhoneNumber = string;
  
  /**
   * This is the Amount transacted normally a numeric value. Money that the customer pays to the Shortcode.
   * Only whole numbers are supported.
   * @type {string}
   * @example "10"
   */
  export type Amount = string;
// ...
  export interface STKPushSuccessfulCallbackBody {
    /**
     * This is the root key for the entire Callback message.
     * @type {Object}
     * @property {Object} stkCallback - This is the first child of the Body.
     */
    Body: {
      stkCallback: {
        /**
         * This is a global unique Identifier for any submitted payment request.
         * This is the same value returned in the acknowledgment message of the initial request.
         * @type {string}
         * @example "29115-34620561-1"
         */
        MerchantRequestID: string;
  
        /**
         * This is a globally unique identifier of the processed checkout transaction request.
         * This is the same value returned in the acknowledgment message of the initial request.
         * @type {string}
         * @example "ws_CO_191220191020363925"
         */
        CheckoutRequestID: string;
  
        /**
         * This is a numeric status code that indicates the status of the transaction processing.
         * 0 means successful processing and any other code means an error occurred or the transaction failed.
         * @type {number}
         * @example 0
         * @example 1032
         */
        ResultCode: number;
  
        /**
         * The result description is a message from the API that gives the status of the request processing.
         * It usually maps to a specific ResultCode value. It can be a Success processing message or an error description message.
         * @type {string}
         * @example "0: The service request is processed successfully."
         * @example "1032: Request canceled by the user"
         */
        ResultDesc: string;
  
        /**
         * This is the JSON object that holds more details for the transaction.
         * It is only returned for successful transactions.
         * @type {Object}
         * @property {Object[]} Item - This is a JSON Array, within the CallbackMetadata, that holds additional transaction details in JSON objects.
         */
        CallbackMetadata: {
          Item: {
            /**
             * This is the Amount that was transacted.
             * @type {number}
             * @example 10500.5
             */
            Amount?: number;
  
            /**
             * This is the unique M-PESA transaction ID for the payment request.
             * The same value is sent to the customer by SMS upon successful processing.
             * @type {string}
             * @example "LHG31AA5TX"
             */
            MpesaReceiptNumber?: string;
  
            /**
             * This is the Balance of the account for the shortcode used as partyB.
             * @type {number}
             * @example 32009.9
             */
            Balance?: number;
  
            /**
             * This is a timestamp that represents the date and time that the transaction was completed in the format YYYYMMDDHHmmss.
             * @type {string}
             * @example "20170827163400"
             */
            TransactionDate?: string;
  
            /**
             * This is the number of the customer who made the payment.
             * @type {string}
             * @example "0722000000"
             */
            PhoneNumber?: string;
          }[];
        };
      };
    };
  }
  
  /**
   * Unsuccessful results body structure.
   * Whenever you receive an error in your callback URL, the unsuccessful transaction will have a body results as below,
   * and the error details will be captured under the Items ResultCode and ResultDesc.
   * You can view a list of possible Lipa na M-PESA online errors on the errors tab.
   * @interface STKPushErrorCallbackBody
   */
  export interface STKPushErrorCallbackBody {
    /**
     * This is the root key for the entire Callback message.
     * @type {Object}
     * @property {Object} stkCallback - This is the first child of the Body.
     */
    Body: {
      stkCallback: {
        /**
         * This is a global unique Identifier for any submitted payment request.
         * @type {string}
         * @example "29115-34620561-1"
         */
        MerchantRequestID: string;
  
        /**
         * This is a globally unique identifier of the processed checkout transaction request.
         * @type {string}
         * @example "ws_CO_191220191020363925"
         */
        CheckoutRequestID: string;
  
        /**
         * This is a numeric status code that indicates the status of the transaction processing.
         * 0 means successful processing and any other code means an error occurred or the transaction failed.
         * @type {number}
         * @example 1032
         */
        ResultCode: number;
  
        /**
         * The result description is a message from the API that gives the status of the request processing.
         * It usually maps to a specific ResultCode value. It can be an error description message.
         * @type {string}
         * @example "Request canceled by user."
         */
        ResultDesc: string;
      };
    };
  }